backend/requirements.txt
# fastapi==0.109.0
# uvicorn==0.27.0
# # psycopg2-binary==2.9.9
# psycopg2-binary==2.9.9 --only-binary=:all:
# langgraph==0.0.20
# langchain==0.1.0
# langchain-google-genai==0.0.6
# google-generativeai==0.3.2
# pydantic==2.5.3
# python-dotenv==1.0.0
# aiohttp==3.9.1
# assemblyai==0.23.0
# elevenlabs==0.2.26
# requests==2.31.0

fastapi
uvicorn
psycopg2-binary
langgraph
langchain
openai
google-generativeai
pydantic
python-dotenv
requests
tqdm
numpy
transformers
sentence-transformers
assemblyai
elevenlabs
aiohttp
gTTS
langchain-google-genai